requirements.txt
# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0

# https://pypi.org/project/google-cloud-storage/
google-cloud-storage==3.0.0